be -> dev

+++ backend/src/main/java/todoktodok/backend/comment/application/service/query/CommentLikeCountDto.java
@@ -11,7 +11,7 @@ public record CommentLikeCountDto(
     ) {
         this(
                 commentId,
-                likeCount.intValue()
+                Math.toIntExact(likeCount)
         );
     }
 }
